Completely decoupled network format parsing (required to properly detect photo-reacions)
