+++ src/jaff/core/parsers/__init__.py
@@ -1,0 +1,4 @@
+from .auxiliary_func import AuxiliaryFunctionParser
+from .network import NetworkParser
+
+__all__ = ["NetworkParser", "AuxiliaryFunctionParser"]
