+++ src/jaff/core/parsers/auxiliary_func/__init__.py
@@ -1,0 +1,3 @@
+from ._engine import AuxiliaryFunctionParser
+
+__all__ = ["AuxiliaryFunctionParser"]
