+++ src/jaff/core/parsers/auxiliary_func/_typing/_functions.py
@@ -2,8 +2,8 @@ from typing import NotRequired, TypedDict
 
 from sympy import Basic
 
-FunctionsDict = TypedDict(
-    "FunctionsDict",
+AuxiliaryFunctionsDict = TypedDict(
+    "AuxiliaryFunctionsDict",
     {
         "def": Basic,
         "args": list[Basic],
