+++ src/jaff/core/parsers/network/__init__.py
@@ -1,0 +1,3 @@
+from ._engine import NetworkParser
+
+__all__ = ["NetworkParser"]
