+++ src/jaff/core/parsers/network/_formats/kida/__init__.py
@@ -1,0 +1,3 @@
+from .reaction import KidaReaction
+
+__all__ = ["KidaReaction"]
