+++ src/jaff/core/parsers/network/_formats/uclchem/__init__.py
@@ -1,0 +1,3 @@
+from .reaction import UclchemReaction
+
+__all__ = ["UclchemReaction"]
