+++ src/jaff/core/parsers/network/_typing/_formats.py
@@ -1,0 +1,21 @@
+from typing import TypedDict
+
+kromeFormatProps = TypedDict(
+    "kromeFormatProps",
+    {
+        "format_nline": int,
+        "idx": bool,
+        "nreact": int,
+        "nprod": int,
+        "tmin": bool,
+        "tmax": bool,
+        "rate": bool,
+    },
+)
+
+prizmoFormatProps = TypedDict(
+    "prizmoFormatProps",
+    {
+        "parse_vars": bool,
+    },
+)
