+++ src/jaff/core/parsers/network/_typing/_parsed.py
@@ -1,0 +1,13 @@
+from typing import TypedDict
+
+parsedListProps = TypedDict(
+    "parsedListProps",
+    {
+        "r": list[str],
+        "p": list[str],
+        "tmin": float | None,
+        "tmax": float | None,
+        "rate": str,
+        "string": str,
+    },
+)
